host connection separated from client
